chore: refactor extract_title_serie_episode_from_metadata
